Created a UI package for menus and returned the player instance to the player package
