+++ ui/button.go
@@ -1,4 +1,4 @@
-package world
+package ui
 
 import (
 	"github.com/hajimehoshi/ebiten/v2"
