+++ world/init.go
@@ -1,18 +1,13 @@
 package world
 
 import (
-	"os"
-
 	"github.com/nath-ellis/AmongSus/space"
 	"github.com/solarlune/resolv"
-	"golang.org/x/image/font"
-	"golang.org/x/image/font/opentype"
 )
 
 var (
-	Speed      float64 = 2
-	MaxSpeed   float64 = 5
-	VCROSDMono font.Face
+	Speed    float64 = 2
+	MaxSpeed float64 = 5
 )
 
 func Init() {
@@ -29,12 +24,4 @@ func Init() {
 	NewObject(1240, 476, "platform")
 
 	space.Space.Add(resolv.NewObject(0, 476, 1200, 124, "object"))
-
-	fontBytes, _ := os.ReadFile("res/text/vcrosdmono.ttf")
-	parsed, _ := opentype.Parse(fontBytes)
-	VCROSDMono, _ = opentype.NewFace(parsed, &opentype.FaceOptions{
-		Size:    76,
-		DPI:     64,
-		Hinting: font.HintingFull,
-	})
 }
